Input Handler Workink
