+++ philosopher.h
@@ -1,0 +1,69 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   philosopher.h                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sueno-te <sueno-te@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/11/11 16:36:43 by sueno-te          #+#    #+#             */
+/*   Updated: 2024/11/11 16:45:54 by sueno-te         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   philo.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: sueno-te <sueno-te@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/11/11 14:37:29 by sueno-te          #+#    #+#             */
+/*   Updated: 2024/11/11 17:08:12 by sueno-te         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef PHILO_H
+# define PHILO_H
+
+# include <unistd.h>
+# include <stdio.h> // fprintf funtion !forbidden to be removed WIP!
+# include <stdlib.h>
+# include <pthread.h>
+# include <sys/time.h>
+
+# define INT_MAX 2147483647
+/*
+    logs 
+• Any state change of a philosopher must be formatted as follows:
+◦ timestamp_in_ms X has taken a fork
+◦ timestamp_in_ms X is eating
+◦ timestamp_in_ms X is sleeping
+◦ timestamp_in_ms X is thinking
+◦ timestamp_in_ms X died
+*/
+
+typedef struct init
+{
+    int number_of_philosophers;
+    int time_to_die;
+    int time_to_eat;
+    int time_to_sleep;
+    int eat_times;
+}t_init;
+
+typedef struct  s_philo
+{
+    int             id;
+    int             eat_count;
+    long long int   last_meal;
+    pthread_t       thread_id;
+    t_init          *holder;
+}               t_philo;
+
+int isspace(int c);
+int ft_atoi(char *str);
+int only_digits(char *str);
+int basic_check(int argc, char *input[]);
+void load_input(t_init *holder, char *input[]);
+
+#endif
